+++ src/app/shared/components/product-card/product-card.component.ts
@@ -1,7 +1,8 @@
+import { AuthService } from './../../../core/services/auth/auth.service';
 import { WishListService } from './../../../core/services/e-comme/wishList/wish-list.service';
 import { Component, inject, Input, OnDestroy, OnInit} from '@angular/core';
 import { Iproduct } from '../../interfaces/product/product';
-import { RouterLink } from '@angular/router';
+import { Router, RouterLink } from '@angular/router';
 import { CartService } from '../../../core/services/e-comme/cart/cart.service';
 import { ToastrService } from 'ngx-toastr';
 import { Subscription } from 'rxjs';
@@ -14,15 +15,17 @@ import { CurrencyPipe, LowerCasePipe, TitleCasePipe } from '@angular/common';
   styleUrl: './product-card.component.scss'
 })
 export class ProductCardComponent  implements OnInit,OnDestroy{
-    // Inject cartService , wishListService and toastrService.
+    // Inject cartService , wishListService , authService , Router and toastrService.
     cartService:CartService=inject(CartService);
     wishListService:WishListService=inject(WishListService);
+    authService:AuthService=inject(AuthService);
+    router:Router=inject(Router);
     toastrService :ToastrService =inject(ToastrService);
     
     // required input of each product.
     @Input({required:true}) product!:Iproduct;
-    // required input of wishList.
-    @Input({required:true}) wishList!:Iproduct[];
+    // input of wishList.
+    @Input() wishList!:Iproduct[];
     
     imgCover!:string;
     heartwishFlag:boolean = false;
@@ -32,11 +35,13 @@ export class ProductCardComponent  implements OnInit,OnDestroy{
   
     //methode to check if the product at wish list or not (Fired at onInit hook).
     checkWishList(){
-      this.wishList.forEach(element => {
-        if(element._id===this.product._id){
-          this.heartwishFlag = true;
-        }
-      });
+      if(this.authService.checkAuthorizedUser()){
+        this.wishList.forEach(element => {
+          if(element._id===this.product._id){
+            this.heartwishFlag = true;
+          }
+        });
+      }
     }
   
     // methode for change image cover
@@ -46,46 +51,59 @@ export class ProductCardComponent  implements OnInit,OnDestroy{
   
     // Add to cart
     addToCart(pId:string){
-      this.addToCartLoading=true;
-      const addToCartSub = this.cartService.addToUserCart(pId).subscribe({
-        next:(res)=>{
-          if(res.status==='success'){
-            this.toastrService.success(`${this.product.title} added successfully to your cart`,'Cart Operations');
-            this.addToCartLoading=false;
-            if(this.heartwishFlag){
-              this.removeFromWishList(pId);
+      if(this.authService.checkAuthorizedUser()){
+        this.addToCartLoading=true;
+        const addToCartSub = this.cartService.addToUserCart(pId).subscribe({
+          next:(res)=>{
+            if(res.status==='success'){
+              this.toastrService.success(`${this.product.title} added successfully to your cart`,'Cart Operations');
+              this.addToCartLoading=false;
+              if(this.heartwishFlag){
+                this.removeFromWishList(pId);
+              }
             }
+          },
+          error:()=>{
+            this.toastrService.error('There is a problem , try again','Cart Operations');
+            this.addToCartLoading=false;
           }
-        },
-        error:()=>{
-          this.toastrService.error('There is a problem , try again','Cart Operations');
-          this.addToCartLoading=false;
-        }
-      });
-      this.subscription.add(addToCartSub);
+        });
+        this.subscription.add(addToCartSub);
+      }
+      else{
+        this.router.navigate(['/logIn']);
+        this.toastrService.error('You are not authorized , please LogIn','Authentication');
+      }
     };
   
     // Add to wishlist
     addToWishList(pId:string){
-      this.updateToWishProductLoading = true;
-      const addToWishListSub = this.wishListService.addToUserWishList(pId).subscribe({
-        next:(res)=>{
-          if(res.status==='success'){
+      if(this.authService.checkAuthorizedUser()){
+        this.updateToWishProductLoading = true;
+        const addToWishListSub = this.wishListService.addToUserWishList(pId).subscribe({
+          next:(res)=>{
+            if(res.status==='success'){
+              this.updateToWishProductLoading = false;
+              this.heartwishFlag = true;
+              this.toastrService.success(`${this.product.title} is added to your wish list`,'Cart Operations');
+            }
+          },
+          error:(err)=>{
             this.updateToWishProductLoading = false;
-            this.heartwishFlag = true;
-            this.toastrService.success(`${this.product.title} is added to your wish list`,'Cart Operations');
+            this.toastrService.error(`There is a problem , try again`,'Cart Operations');
           }
-        },
-        error:(err)=>{
-          this.updateToWishProductLoading = false;
-          this.toastrService.error(`There is a problem , try again`,'Cart Operations');
-        }
-      });
-      this.subscription.add(addToWishListSub);
+        });
+        this.subscription.add(addToWishListSub);
+      }
+      else{
+        this.router.navigate(['/logIn']);
+        this.toastrService.error('You are not authorized , please LogIn','Authentication');
+      } 
     }
   
     // Remove from wishList method.
     removeFromWishList(pId:string){
+      if(this.authService.checkAuthorizedUser()){
         this.updateToWishProductLoading = true;
         const removeWishListSub = this.wishListService.removeFromUserWishList(pId).subscribe({
           next:(res)=>{
@@ -100,7 +118,12 @@ export class ProductCardComponent  implements OnInit,OnDestroy{
             this.toastrService.error(`There is a problem , try again`,'Cart Operations');
           }
         });
-        this.subscription.add(removeWishListSub);
+        this.subscription.add(removeWishListSub); 
+      } 
+      else{
+        this.router.navigate(['/logIn']);
+        this.toastrService.error('You are not authorized , please LogIn','Authentication');
+      }
     }  
   
     ngOnInit(): void {
